backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
neo4j>=5.17.0
langchain>=0.1.10
langchain-community>=0.0.25
langchain-google-genai
langgraph
gitpython>=3.1.41
tree-sitter>=0.20.4
tree-sitter-languages>=1.10.2
httpx>=0.27.0
tree-sitter-python
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
python-multipart
PyPDF2>=3.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
